Splitting up NVIM config
