--- nvim/lua/custom/plugins/init.lua
@@ -1,5 +1,0 @@
--- You can add your own plugins here or in other files in this directory!
---  I promise not to create any merge conflicts in this directory :)
---
--- See the kickstart.nvim README for more information
-return {}
