--- nvim/lua/kickstart/plugins/autopairs.lua
@@ -1,8 +1,0 @@
--- autopairs
--- https://github.com/windwp/nvim-autopairs
-
-return {
-  'windwp/nvim-autopairs',
-  event = 'InsertEnter',
-  opts = {},
-}
