+++ nvim/lua/nerevar/init.lua
@@ -1,0 +1,2 @@
+require 'nerevar.base'
+require 'nerevar.lazy'
