+++ nvim/lua/plugins/gitsigns.lua
@@ -1,0 +1,4 @@
+return { -- Adds git related signs to the gutter, as well as utilities for managing changes
+    'lewis6991/gitsigns.nvim',
+    lazy = false,
+}
