+++ nvim/lua/plugins/todo-comments.lua
@@ -1,0 +1,10 @@
+return { 
+    'folke/todo-comments.nvim', 
+    event = 'VimEnter', 
+    dependencies = { 
+        'nvim-lua/plenary.nvim' 
+    }, 
+    opts = { 
+        signs = false 
+    } 
+}
